ondevice-ai/python_assistant/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "privacy-assistant"
version = "0.1.0"
description = "Privacy-first AI assistant core built on FastAPI with MLX inference."
requires-python = ">=3.10"
authors = [
  { name = "MahiLLM" }
]
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "cryptography>=42.0.0",
  "keyring>=24.3.0",
  "numpy>=1.26.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.20.0",
  "langchain>=0.3.0",
  "typer>=0.12.3",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
ml = [
  "mlx>=0.12.0; platform_system == 'Darwin'",
  "mlx-lm>=0.5.1; platform_system == 'Darwin'",
  "sentence-transformers>=3.0.0",
  "faiss-cpu>=1.8.0"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.privacy_assistant]
default_model = "mlx-community/Meta-Llama-3-8B-Instruct"
vector_db_path = "./var/pki.sqlite"
encryption_key_service = "privacy-assistant"
